third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/iframe-expected.txt

This is a testharness.js-based test.
PASS Same-host fetches are strictly same-site
PASS Subdomain fetches are strictly same-site
FAIL Cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.19127505789765564" but got (undefined) undefined
PASS Same-host redirecting to same-host fetches are strictly same-site
PASS Subdomain redirecting to same-host fetches are strictly same-site
PASS Cross-site redirecting to same-host fetches are strictly same-site
PASS Same-host redirecting to subdomain fetches are strictly same-site
PASS Subdomain redirecting to subdomain fetches are strictly same-site
PASS Cross-site redirecting to subdomain fetches are strictly same-site
FAIL Same-host redirecting to cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.29605694718703246" but got (undefined) undefined
FAIL Subdomain redirecting to cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.02810112070770998" but got (undefined) undefined
FAIL Cross-site redirecting to cross-site fetches are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.5694235788970805" but got (undefined) undefined
Harness: the test ran to completion.
